+++ src/main/java/gift/member/exception/InvalidKakaoTalkTemplateException.java
@@ -1,4 +1,4 @@
-package gift.kakao.exception;
+package gift.member.exception;
 
 public class InvalidKakaoTalkTemplateException extends RuntimeException {
