+++ src/main/java/gift/member/repository/KakaoTokenRepository.java
@@ -1,6 +1,6 @@
-package gift.kakao.repository;
+package gift.member.repository;
 
-import gift.kakao.entity.KakaoToken;
+import gift.member.entity.KakaoToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KakaoTokenRepository extends JpaRepository<KakaoToken, String> {
